* commit '6e38b69cdd38d10684f935900ba27b684c90fab0':
  [EDELIVERY-8424] [EDELIVERY-8237] Fix long int overflow for entity id passed to the UI.  [part of PR #2476]
  [EDELIVERY-8424] change more entityIds from int to long
  [EDELIVERY-8424] fix unit test
  [EDELIVERY-8424] forgotten pmode int ids
  [EDELIVERY-8424] [EDELIVERY-8336] put long instead of int for pMode id [PR #2571]
  [EDELIVERY-8424] [EDELIVERY-8239] Fix issue with pMode not showing in the UI [PR #2520]
  [EDELIVERY-8424] [EDELIVERY-8334] fix command propagation on oracle [PR #2553]

+++ Domibus-MSH-ext-services-delegate/src/main/java/eu/domibus/ext/delegate/services/pmode/PModeServiceDelegate.java
@@ -37,7 +37,7 @@ public class PModeServiceDelegate implements PModeExtService {
     MultiPartFileUtil multiPartFileUtil;
 
     @Override
-    public byte[] getPModeFile(int id) {
+    public byte[] getPModeFile(long id) {
         return pModeService.getPModeFile(id);
     }
